% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusLabelFree.R
\docType{data}
\name{annotations.id}
\alias{annotations.id}
\title{Protein annotations for merging into \code{proteusData} object}
\format{A data frame with 3796 rows and 4 columns}
\description{
Protein annotations formatted for \code{annotateProteins} function.
}
