% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusLabelFree.R
\docType{data}
\name{annotations}
\alias{annotations}
\title{Protein annotations}
\format{A data frame with 9067 rows and 3 columns}
\description{
Protein annotations obtained with \code{fetchFromUniProt}.
}
