% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusLabelFree.R
\docType{data}
\name{evi}
\alias{evi}
\title{Evidence data}
\format{A data frame with 470318 rows and 10 columns}
\description{
Evidence data from a proteomics experiment in two conditions, 7 replicates each.
}
