% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusUnlabelled.R
\docType{data}
\name{meta}
\alias{meta}
\title{Proteomics metadata}
\format{A \code{data.frame} with 14 rows and 3 columns (sample, condition and replicate).}
\description{
Metadata describing a proteomics experiment in two conditions with 7 replicates each.
}
