% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusUnlabelled.R
\docType{data}
\name{pepdat.clean}
\alias{pepdat.clean}
\title{Clean peptide data}
\format{A \code{proteusData} object}
\description{
Peptide data aggregated from evidence data \code{evi}. One "bad" replicate was removed.
}
