% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusUnlabelled.R
\docType{data}
\name{proteusUnlabelled}
\alias{proteusUnlabelled}
\title{Proteus example unlabelled data}
\format{Six objects from different stages of data processing:
\describe{
  \item{\code{evi}}{Pre-processed evidence data (data frame)}
  \item{\code{meta}}{Metadata (data frame)}
  \item{\code{meta.clean}}{Clean metadata (data frame)}
  \item{\code{pepdat}}{Peptide data (\code{proteusData} object)}
  \item{\code{pepdat.clean}}{Clean peptide data (\code{proteusData} object)}
  \item{\code{prodat}}{Protein data (\code{proteusData} object)}
  \item{\code{annotations}}{Protein annotations downloaded from UniProt}
}
This package also contains the full original evidence file, peptides file, proteinGroups
file and formatted metadata file (see examples below).}
\description{
Example data set for \code{proteus} package. It comes from a label-free MS/MS
experiment in two conditions (called 1112 and BMO-1) and seven replicates each.
}
\examples{
data(proteusUnlabelled)
meta
head(pepdat$tab)
head(prodat$tab)

evidenceFile <- system.file("extdata", "evidence.txt.gz", package="proteusUnlabelled")
peptidesFile <- system.file("extdata", "peptides.txt.gz", package="proteusUnlabelled")
proteinGroupsFile <- system.file("extdata", "proteinGroups.txt.gz", package="proteusUnlabelled")
metadataFile <- system.file("extdata", "metadata.txt", package="proteusUnlabelled")
}
